Doing stuff

+++ app/src/main/java/com/example/superawesometodolistcatgaggingappgagging/api/TodoFetchWorker.kt
@@ -2,13 +2,11 @@ package com.example.superawesometodolistcatgaggingappgagging.api
 
 import android.content.Context
 import android.util.Log
-import androidx.lifecycle.ViewModelProvider.AndroidViewModelFactory
-import androidx.lifecycle.viewmodel.CreationExtras
 import androidx.work.CoroutineWorker
 import androidx.work.WorkerParameters
 import com.example.superawesometodolistcatgaggingappgagging.database.TodoTable
-import com.example.superawesometodolistcatgaggingappgagging.database.TodosApplication
 import com.example.superawesometodolistcatgaggingappgagging.database.TodosDatabase
+import net.gotev.cookiestore.SharedPreferencesCookieStore
 import okhttp3.JavaNetCookieJar
 import okhttp3.OkHttpClient
 import okhttp3.RequestBody
@@ -19,28 +17,38 @@ import retrofit2.http.Body
 import retrofit2.http.GET
 import retrofit2.http.POST
 import java.net.CookieManager
+import java.net.CookiePolicy
 
 
 private const val BASE_URL = "https://satdlcgag.saiger.dev/"
-
-private val interceptor = HttpLoggingInterceptor().setLevel(HttpLoggingInterceptor.Level.BODY)
-// Used to persist the session, meaning if we login, we stay logged in until we close the app
-private val cookieHandler = CookieManager()
-private val client = OkHttpClient.Builder()
-    .addInterceptor(interceptor)
-    .cookieJar(JavaNetCookieJar(cookieHandler))
-    .build()
-
-private val retrofit = Retrofit.Builder()
-    .addConverterFactory(GsonConverterFactory.create())
-    .baseUrl(BASE_URL)
-    .client(client)
-    .build()
-
-
 object TodoApi {
-    val retrofitService: TodoApiService by lazy {
-        retrofit.create(TodoApiService::class.java)
+    var cookieStore: SharedPreferencesCookieStore? = null
+    var retrofitService: TodoApiService? = null;
+
+    fun init(context: Context) {
+        if (retrofitService == null) {
+            cookieStore = SharedPreferencesCookieStore(context.applicationContext, "loginCookie")
+
+            val interceptor = HttpLoggingInterceptor().setLevel(HttpLoggingInterceptor.Level.BODY)
+
+            // Used to persist the session
+            val cookieHandler = CookieManager(
+                cookieStore,
+                CookiePolicy.ACCEPT_ORIGINAL_SERVER
+            )
+            val client = OkHttpClient.Builder()
+                .addInterceptor(interceptor)
+                .cookieJar(JavaNetCookieJar(cookieHandler))
+                .build()
+
+            val retrofit = Retrofit.Builder()
+                .addConverterFactory(GsonConverterFactory.create())
+                .baseUrl(BASE_URL)
+                .client(client)
+                .build()
+
+            retrofitService = retrofit.create(TodoApiService::class.java)
+        }
     }
 }
 
@@ -68,21 +76,24 @@ class TodoFetchWorker(
     context: Context,
     workerParams: WorkerParameters
 ) : CoroutineWorker(context, workerParams) {
-
     override suspend fun doWork(): Result {
         val db = TodosDatabase.getDatabase(applicationContext)
         val dao = db.todoDao()
         return try {
-            val response = TodoApi.retrofitService.fetchTodoList()
-            if (response.success == 1) {
-                Log.d("Worker", "Fetched ${response.data.todos.size} todos")
-                dao.deleteAll()
-                dao.insertAll(*response.data.todos.map {
-                    TodoTable(it.id, it.name, it.desc, it.time)
-                }.toTypedArray())
-                Result.success()
+            if (TodoApi.retrofitService != null) {
+                val response = TodoApi.retrofitService!!.fetchTodoList()
+                if (response.success == 1) {
+                    Log.d("Worker", "Fetched ${response.data.todos.size} todos")
+                    dao.deleteAll()
+                    dao.insertAll(*response.data.todos.map {
+                        TodoTable(it.id, it.name, it.desc, it.time)
+                    }.toTypedArray())
+                    Result.success()
+                } else {
+                    Log.e("Worker", "API error")
+                    Result.retry()
+                }
             } else {
-                Log.e("Worker", "API error")
                 Result.retry()
             }
         } catch (e: Exception) {
